[pf-476] Remove useless code and module (scoring-formulas)
